+++ ProjectSettings/ProjectVersion.txt
@@ -1,0 +1,2 @@
+m_EditorVersion: 5.3.2f1
+m_StandardAssetsVersion: 0
